[TASK] Get rid of the AbstractController
It is confusing to ppl, why our controllers do not extend the Actioncontroller

Releases: main, 13.4
